Cambios en el tipo de datos (DNIs y Telefonos

+++ Turnos/Models/paciente.cs
@@ -23,8 +23,8 @@ namespace Turnos.Models
         public int idPaciente { get; set; }
         public string nombre { get; set; }
         public string apellido { get; set; }
-        public decimal dni { get; set; }
-        public Nullable<decimal> telefono { get; set; }
+        public int dni { get; set; }
+        public Nullable<int> telefono { get; set; }
         public string mail { get; set; }
         public int idObraSocial { get; set; }
         public string sexo { get; set; }
